:white_check_mark: Add test damage with rage

+++ Assets/Modules/Hero/Scripts/Warrior.cs
@@ -29,7 +29,7 @@ namespace Aloha.Hero
 
         public override void Attack(Entity entity)
         {
-            int damage = this.stats.attack * ATTACK_BONUS;
+            int damage = (int) (this.stats.attack * ATTACK_BONUS);
             entity.TakeDamage(damage);
             currentRage = currentRage / 2;
         }
